requirements.txt
torch>=2.0
transformers>=4.40
datasets
bitsandbytes
accelerate
pydantic>=2.0
pyyaml
tqdm
numpy
pandas
matplotlib
seaborn
scikit-learn
